Rename plugmngr to pkgmngr
